@@ -300,13 +302,13 @@
     "typescript-plugin-styled-components": "1.0.0",
     "typescript-styled-plugin": "0.13.0",
     "typings-for-css-modules-loader": "1.7.0",
-    "uglifyjs-webpack-plugin": "2.1.0",
+    "uglifyjs-webpack-plugin": "2.1.1",
     "url-loader": "1.1.2",
-    "webpack": "4.27.1",
+    "webpack": "4.28.3",
     "webpack-bundle-analyzer": "3.0.3",
     "webpack-chunk-hash": "0.6.0",
     "webpack-cli": "3.1.2",
-    "webpack-dev-server": "3.1.13",
+    "webpack-dev-server": "3.1.14",
     "webpack-manifest-plugin": "2.0.4",
     "webpack-merge": "4.1.5",
     "webpack-notifier": "1.7.0",
