@@ -1,18 +1,20 @@
-// import * as React from 'react'
-// import { renderToString } from 'react-dom/server'
-// // import { Provider } from 'react-redux'
-// // import configureStore from './redux/configureStore'
-// import { App } from './app'
+import * as React from 'react'
+import { renderToString } from 'react-dom/server'
+// import { Provider } from 'react-redux'
+// import configureStore from './redux/configureStore'
+import App from './app'
 
-// // module.exports = function render(initialState) {
-// export const render = () => {
-//   // const store = configureStore(initialState)
-//   const content = renderToString(
-//     // <Provider store={store}>
-//     <App />
-//     // </Provider>
-//   )
+export const render = (initialState: any) => {
+  // const store = configureStore(initialState)
+  // <Provider store={store}></Provider>
 
-//     // const preloadedState = store.getState()
-//     // return { content, preloadedState }
-// }
+  console.log(initialState)
+
+  const content = renderToString(
+    <App />
+  )
+
+  // const preloadedState = store.getState()
+  // return { content, preloadedState }
+  return { content, {} }
+}
