@@ -1,9 +1,21 @@
 import * as React from 'react'
 import { render } from 'react-dom'
-import App from './app'
+import { Provider } from 'mobx-react'
 // import * as serviceWorker from './serviceWorker'
 
-import { environment } from 'settings/environment'
+import { RouterStore, syncHistoryWithStore } from 'mobx-react-router'
+
+import App from './app'
+
+import { createStore } from 'store'
+import { environment } from 'settings'
+import { browserHistory } from 'helpers'
+
+const routingStore = new RouterStore()
+
+const store = createStore(routingStore)
+
+const history = syncHistoryWithStore(browserHistory, routingStore)
 
 const appRoot = document.getElementById('app')
 
@@ -12,7 +24,12 @@ if (appRoot == null) {
 }
 
 const renderApp = () => {
-  render(<App />, appRoot)
+  render(
+    <Provider {...store}>
+      <App history={history} />
+    </Provider>,
+    appRoot
+  )
   document.body.classList.remove('loading')
 }
 
