@@ -1,7 +1,7 @@
 import store from 'store2'
 
-export const TOKEN = {
+export const AccessToken = {
   set: (value: string) => store.set('token', value),
   get: () => store.get('token'),
-  del: () => store.remove('token'),
+  clear: () => store.remove('token'),
 }
