@@ -1,7 +1,7 @@
 export * from './api'
 export * from './classes'
 export * from './history'
-export * from './token'
+export * from './access-token'
 export * from './utils/range'
 export * from './utils/preload'
 export * from './utils/animation'
