@@ -1,32 +1,28 @@
 import axios from 'axios'
 
-import { TOKEN } from '../token'
+import { AccessToken } from '../access-token'
 import { API_URL } from '../api'
 
-const headers: any = {
-  'Accept': 'application/json',
-  'Content-Type': 'application/json',
-}
+export const refreshToken = () => ({})
 
-export const request: any = (options?: any = {}): any => {
+export const request: any = (options: any = { noToken: false }) => {
   const noToken = Object.keys(options).length && options.hasOwnProperty('noToken') && options.noToken
 
-  if (!noToken) {
-    const authToken = TOKEN.get()
+  const headers: any = {
+    'Accept': 'application/json',
+    'Content-Type': 'application/json',
+  }
 
-    if (authToken) {
-      headers.Authorization = `Token ${authToken}`
+  if (!noToken) {
+    const authorizationToken = AccessToken.get()
+
+    if (authorizationToken) {
+      headers['x-access-token'] = `${authorizationToken}`
     }
   }
 
-  const config: any = {
+  return axios.create({
     baseURL: API_URL,
     headers,
-  }
-
-  if (options) {}
-
-  const instance = axios.create(config)
-
-  return instance
+  })
 }
