@@ -1,11 +1,11 @@
 declare module '*.svg' {
-    const content: any;
-    export default content;
+  const content: any;
+  export default content;
 }
 
 declare module '*.png' {
-    const content: any;
-    export default content;
+  const content: any;
+  export default content;
 }
 
 declare module '*.jpg';
