@@ -1,36 +1,27 @@
 const define = require('./define')
 
 module.exports = {
-    nodeEnv: define.rs_mode,
-    minimize: true,
-    concatenateModules: true,
-    runtimeChunk: {
-        name: 'startup',
+  minimize: define.rs_release,
+  chunkIds: 'named',
+  nodeEnv: define.rs_mode,
+  concatenateModules: true,
+  runtimeChunk: {
+    name: 'startup',
+  },
+  splitChunks: {
+    cacheGroups: {
+      default: {
+        minChunks: 2,
+        priority: -20,
+        reuseExistingChunk: true,
+      },
+      bundle: {
+        test: /[\\/]node_modules[\\/]/,
+        chunks: 'all',
+        name: 'bundle',
+        priority: -10,
+        enforce: true,
+      },
     },
-    splitChunks: {
-        chunks: 'async',
-        minSize: 30000,
-        maxSize: 0,
-        minChunks: 1,
-        maxAsyncRequests: 5,
-        maxInitialRequests: 3,
-        automaticNameDelimiter: '~',
-        name: true,
-        cacheGroups: {
-            bundle: {
-                test: /[\\/]node_modules[\\/]/,
-                name: 'bundle',
-                minChunks: 1,
-                priority: -10,
-                chunks: 'initial',
-                enforce: true,
-                reuseExistingChunk: true,
-            },
-            default: {
-                minChunks: 2,
-                priority: -20,
-                reuseExistingChunk: true,
-            },
-        },
-    },
+  },
 }
